extract e2e utils
